+++ ui/assets/useUserAgents-B3XGFWz9.js
@@ -1,0 +1,1 @@
+import{u as r,Z as s,Q as t}from"./index-pazFDlPV.js";function n(e={}){return r(t.userAgents(),()=>s.getAgents(),{pollInterval:e.pollInterval})}export{n as u};